requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
streamlit==1.31.1
requests==2.31.0

# Vector store
qdrant-client==1.7.3

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1

# Environment variables
python-dotenv==1.0.1

# Type hints
typing-extensions==4.9.0

# Frontend
streamlit==1.31.1

# Local models
ollama==0.1.6

# Utilities
numpy==1.26.4
pytest==8.0.2